Add logical operators sample to demonstrate short-circuiting

+++ Decimals/Program.cs
@@ -6,13 +6,13 @@ namespace Decimals
     {
         static void Main(string[] args)
         {
-            /*for (double x = 0; x != 10.0; x += 0.2)
+            /*for (float x = 0; x != 10.0; x += 0.2)
             {
                 Console.Write(x);
                 Console.ReadLine();
             }*/
 
-            for (double x = 0; x != 10.0; x += 0.2)
+            for (float x = 0; x != 10.0; x += 0.2f)
             {
                 Console.Write(DoubleConverter.ToExactString(x));
                 Console.ReadLine();
